file restructure and model installer
